requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
asynch==0.2.1
asyncio==3.4.3
asyncpg==0.25.0
certifi==2022.5.18.1
ciso8601==2.2.0
click==8.1.3
clickhouse-cityhash==1.0.2.3
clickhouse-driver==0.2.4
fastapi==0.85.0
h11==0.13.0
idna==3.3
leb128==1.0.4
lz4==4.0.1
Mako==1.2.0
MarkupSafe==2.1.1
prometheus-client==0.14.1
prometheus-fastapi-instrumentator==5.8.2
psycopg2==2.9.3
pydantic==1.9.1
pyhumps==3.5.3
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
sentry-sdk==1.9.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.20.4
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.17.6
zstd==1.5.2.5
